add plot module
