+++ CPPN/__init__.py
@@ -1,2 +1,6 @@
-from color_factory import ColorFactory
-from network import NN
+from .network import NN
+from .color_factory import ColorFactory
+
+from .plot_tools import plot_input
+from .plot_tools import plot_image
+from .plot_tools import plot_pattern
